set and cancel reservation done

+++ proj2/kernel/rsv.h
@@ -3,5 +3,5 @@
 #include <linux/sched.h>
 #include <linux/hrtimer.h>
 
-asmlinkage int set_rsv(pid_t pid, struct timespec *C, struct timespec *T);
-asmlinkage int cancel_rsv(pid_t pid);
+asmlinkage long set_rsv(pid_t pid, struct timespec *C, struct timespec *T);
+asmlinkage long cancel_rsv(pid_t pid);
